backend/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
psycopg2-binary==2.9.1
gunicorn==20.1.0
requests==2.28.1
djangorestframework-simplejwt===5.2.0
